requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
python-multipart>=0.0.20
google-genai>=1.4.0
pydantic-settings>=2.7.0
python-dotenv>=1.0.0
pytest>=8.0.0
httpx>=0.28.0
ruff>=0.9.0
